split output component
