+++ optuna_dashboard/ts/components/PreferentialTrials.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useEffect, useState, useMemo } from "react"
+import React, { FC, useEffect, useState } from "react"
 import {
   Typography,
   Box,
@@ -32,8 +32,7 @@ import {
   isThreejsArtifact,
   useThreejsArtifactModal,
 } from "./ThreejsArtifactViewer"
-import { ArtifactCardMedia } from "./ArtifactCardMedia"
-import { MarkdownRenderer } from "./Note"
+import { PreferentialOutputComponent } from "./PreferentialOutputComponent"
 
 const SettingsPage: FC<{
   studyDetail: StudyDetail
@@ -167,29 +166,6 @@ const isComparisonReady = (
     return artifact !== undefined
   }
   return false
-}
-
-export const OutputContent: FC<{
-  trial: Trial
-  artifact?: Artifact
-  componentType: FeedbackComponentType
-  urlPath: string
-}> = ({ trial, artifact, componentType, urlPath }) => {
-  const note = useMemo(() => {
-    return <MarkdownRenderer body={trial.note.body} />
-  }, [trial.note.body])
-  if (componentType === undefined || componentType.output_type === "note") {
-    return note
-  }
-  if (componentType.output_type === "artifact") {
-    if (artifact === undefined) {
-      return null
-    }
-    return (
-      <ArtifactCardMedia artifact={artifact} urlPath={urlPath} height="100%" />
-    )
-  }
-  return null
 }
 
 export const getArtifactUrlPath = (
@@ -338,7 +314,7 @@ const PreferentialTrial: FC<{
       >
         {isReady ? (
           <>
-            <OutputContent
+            <PreferentialOutputComponent
               trial={trial}
               artifact={artifact}
               componentType={componentType}
